api/FAQCategory/views.py
from django.shortcuts import render
from django.db.models import Q

from rest_framework.permissions import IsAuthenticated, AllowAny
from rest_framework.response import Response
from rest_framework.decorators import action
from rest_framework.filters import SearchFilter, OrderingFilter
from rest_framework import viewsets, status
from rest_framework_extensions.mixins import NestedViewSetMixin

from django_filters.rest_framework import DjangoFilterBackend

from FAQCategory.models import (
    FAQCategory
)

from FAQCategory.serializers import (
    FAQCategorySerializer,
    FAQCategoryHistorySerializer
)

class FAQCategoryViewSet(NestedViewSetMixin, viewsets.ModelViewSet):
    queryset = FAQCategory.objects.all()
    serializer_class = FAQCategorySerializer
    filter_backends = (DjangoFilterBackend, SearchFilter, OrderingFilter)
    filterset_fields = [
        # 'FAQId', 
        'active',
        'categoryId',
        # 'title',
        # 'Description',
        # 'ProductRegNo',
        # 'createdBy',
        # 'created_date',
    ]

    def get_permissions(self):
        if self.action == 'list':
            permission_classes = [AllowAny]
        else:
            permission_classes = [AllowAny]

        return [permission() for permission in permission_classes]    

    
    def get_queryset(self):
        queryset = FAQCategory.objects.all()
        # queryset = CustomUser.objects.filter(string__contains=CustomUser.name)

        """
        if self.request.user.is_anonymous:
            queryset = Company.objects.none()

        else:
            user = self.request.user
            company_employee = CompanyEmployee.objects.filter(employee=user)
            company = company_employee[0].company
            
            if company.company_type == 'AD':
                queryset = User.objects.all()
            else:
                queryset = User.objects.filter(company=company.id)
        """
        return queryset   
    
    @action(methods=['GET','POST'], detail=True)
    def history(self, request, *args, **kwargs):
        history = FAQCategory.history.all()
        serializer = FAQCategorySerializer(history, many=True)
        return Response(serializer.data)
    
class FAQCategoryHistoryViewSet(NestedViewSetMixin, viewsets.ModelViewSet):
    queryset = FAQCategory.history.all()
    serializer_class = FAQCategoryHistorySerializer
    filter_backends = (DjangoFilterBackend, SearchFilter, OrderingFilter)
    # filterset_fields = []

    def get_permissions(self):
        permission_classes = [AllowAny]

        return [permission() for permission in permission_classes]    

    
    def get_queryset(self):
        queryset = FAQCategory.history.all()
        return queryset

